backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
redis>=5.0.1
google-search-results>=2.4.2
httpx>=0.28.0
loguru>=0.7.0
